backend/requirements.txt
flask==1.0.2
newspaper3k
aylien-apiclient
ibm-watson>=3.0.3
